// frontend/src/app/models/donhang.model.ts
export interface DonHang {
    id: number;
    ten: string;
    diaChi: string;
    sdt: string;
    ngayDat: Date;
    kieuGiaoHang: number;
    ghiChu: string;
    trangThai: number;
    idPhuongThuc: number;
    idNguoiDung: number;
    tongTien: number;
    tenPhuongThuc: string;
    taiKhoan: string;
}